ragchecker

+++ modules/retrieve.py
@@ -158,7 +158,7 @@ class Retrieve:
         num_emb = 0
         for emb_chunk in doc_embeds:
             emb_chunk = emb_chunk.to('cuda')
-            scores_q = self.model.similarity_fn(emb_q, emb_chunk)
+            scores_q = self.model.similarity_fn(emb_q.float(), emb_chunk.float())
             # if detach_and_cpu:
             #     scores_q = scores_q.detach().cpu().float()
             scores_sorted_q, indices_sorted_q = torch.topk(scores_q, top_k_documents, dim=1)
